Fixed keyreleasehandler, added bypass to urlcheck
Checks if the keyrelease event has a key property Alt on key release
instead of checking if it has the altKey property set to true,
which is a more uniform way to check for alt key press and release.

Removed keypressHandler and the keydown listener as the intention is to
only temporarily skip deplaylistifying when the alt key is tapped.
This makes it much easier to manage and gets us out of having to do
complex async messaging and handling timing based edgecases.

Bypass deplaylistifying for 5s on alt key release by checking for the
string variable skip_checking to be 'true' in local storage. This lets
us bypass deplaylistifying across multiple tabs running YouTube and is a
comfortable amount of time for users to click on links after activating
the bypass.

Add icon attributions to README